+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Function.call_times
 - Function.once
 - Delegation functions in order to use `Exdash.<function>` instead of `Exdash.<module>.<function>`
+- Function.curry
 
 ### Change
 - Function.after_nth doctest